Mutability (#419)
* mutability

* hack

* rename Map into StringMap

* rename

* modules

* immutableMap

* asyncSet, asyncGet.

* IO

* deno
