+++ djs/transpiler/module.f.ts
@@ -3,16 +3,16 @@ import { type Result, error, ok } from '../../types/result/module.f.ts'
 import { fold, drop, map as listMap, type List, toArray, includes } from '../../types/list/module.f.ts'
 import type * as Operator from '../../types/function/operator/module.f.ts'
 import { tokenize } from '../tokenizer/module.f.ts'
-import { setReplace, at, type Map } from '../../types/map/module.f.ts'
-import type { Fs, BufferEncoding } from '../../io/module.f.ts'
+import { setReplace, at, type OrderedMap } from '../../types/ordered_map/module.f.ts'
+import type { Fs } from '../../io/module.f.ts'
 import { stringToList } from '../../text/utf16/module.f.ts'
 import { concat as pathConcat } from '../../path/module.f.ts'
 import { parseFromTokens } from '../parser/module.f.ts'
 import { run, type AstModule } from '../ast/module.f.ts'
 
-export type ParseContext = {    
-    readonly fs: Fs    
-    readonly complete: Map<djsResult>
+export type ParseContext = {
+    readonly fs: Fs
+    readonly complete: OrderedMap<djsResult>
     readonly stack: List<string>
     readonly error: string | null
 }
@@ -54,10 +54,10 @@ const parseModule
     = path => context => {
         const content = context.fs.readFileSync(path, 'utf8')
         if (content === null) {
-            return error('file not found')  
+            return error('file not found')
         }
 
-        const tokens = tokenize(stringToList(content))        
+        const tokens = tokenize(stringToList(content))
         return parseFromTokens(tokens)
 }
